+++ st2actions/st2actions/cmd/actionrunner.py
@@ -4,15 +4,15 @@ import sys
 
 from oslo.config import cfg
 
+from st2actions import config
+from st2actions import scheduler, worker
 from st2common import log as logging
+from st2common.constants.logging import DEFAULT_LOGGING_CONF_PATH
 from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
-from st2common.constants.logging import DEFAULT_LOGGING_CONF_PATH
-from st2common.transport.utils import register_exchanges
 from st2common.signal_handlers import register_common_signal_handlers
-from st2actions import config
-from st2actions import worker
-
+from st2common.transport.utils import register_exchanges
+from st2common.triggers import register_internal_trigger_types
 
 LOG = logging.getLogger(__name__)
 
@@ -46,20 +46,41 @@ def _setup():
     register_common_signal_handlers()
 
     # 4. Register internal triggers
-    # Note: We need to do import here because of a messed up configuration
-    # situation (this module depends on configuration being parsed)
-    from st2common.triggers import register_internal_trigger_types
     register_internal_trigger_types()
 
 
 def _run_worker():
     LOG.info('(PID=%s) Worker started.', os.getpid())
+
+    components = [
+        scheduler.get_scheduler(),
+        worker.get_worker()
+    ]
+
     try:
-        worker.work()
+        for component in components:
+            component.start()
+
+        for component in components:
+            component.wait()
     except (KeyboardInterrupt, SystemExit):
         LOG.info('(PID=%s) Worker stopped.', os.getpid())
+
+        errors = False
+
+        for component in components:
+            try:
+                component.shutdown()
+            except:
+                LOG.exception('Unable to shutdown %s.', component.__class__.__name__)
+                errors = True
+
+        if errors:
+            return 1
     except:
+        LOG.exception('(PID=%s) Worker unexpectedly stopped.', os.getpid())
         return 1
+
     return 0