+++ st2actions/st2actions/notifier.py
@@ -13,18 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import eventlet
+import json
+
 from kombu import Connection
-from kombu.mixins import ConsumerMixin
 from oslo.config import cfg
 
 from st2common import log as logging
 from st2common.constants.action import LIVEACTION_STATUS_SUCCEEDED, LIVEACTION_STATUS_FAILED
 from st2common.constants.triggers import INTERNAL_TRIGGER_TYPES
+from st2common.models.db import action as action_models
 from st2common.models.system.common import ResourceReference
-from st2common.transport import liveaction, publishers
+from st2common.transport import consumers, liveaction, publishers
 from st2common.transport.reactor import TriggerDispatcher
-from st2common.util.greenpooldispatch import BufferedDispatcher
+
 
 LOG = logging.getLogger(__name__)
 
@@ -37,46 +38,11 @@ ACTION_TRIGGER_TYPE = INTERNAL_TRIGGER_TYPES['action'][0]
 NOTIFY_TRIGGER_TYPE = INTERNAL_TRIGGER_TYPES['action'][1]
 
 
-class LiveActionUpdateQueueConsumer(ConsumerMixin):
-    def __init__(self, connection, notifier):
-        self.connection = connection
-        self._dispatcher = BufferedDispatcher()
-        self._notifier = notifier
+class Notifier(consumers.MessageHandler):
+    message_type = action_models.LiveActionDB
 
-    def shutdown(self):
-        self._dispatcher.shutdown()
-
-    def get_consumers(self, Consumer, channel):
-        consumer = Consumer(queues=[ACTIONUPDATE_WORK_Q],
-                            accept=['pickle'],
-                            callbacks=[self.process_task])
-        # use prefetch_count=1 for fair dispatch. This way workers that finish an item get the next
-        # task and the work does not get queued behind any single large item.
-        consumer.qos(prefetch_count=1)
-        return [consumer]
-
-    def process_task(self, body, message):
-        LOG.debug('process_task')
-        LOG.debug('     body: %s', body)
-        LOG.debug('     message.properties: %s', message.properties)
-        LOG.debug('     message.delivery_info: %s', message.delivery_info)
-        try:
-            self._dispatcher.dispatch(self._do_process_task, body)
-        finally:
-            message.ack()
-
-    def _do_process_task(self, body):
-        try:
-            if body.status in ACTION_COMPLETE_STATES:
-                self._notifier.handle_action_complete(body)
-        except:
-            LOG.exception('Sending notifications/action trigger failed. Message body : %s', body)
-
-
-class Notifier(object):
-    def __init__(self, q_connection=None, trigger_dispatcher=None):
-        self._queue_consumer = LiveActionUpdateQueueConsumer(q_connection, self)
-        self._consumer_thread = None
+    def __init__(self, connection, queues, trigger_dispatcher=None):
+        super(Notifier, self).__init__(connection, queues)
         self._trigger_dispatcher = trigger_dispatcher
         self._notify_trigger = ResourceReference.to_string_reference(
             pack=NOTIFY_TRIGGER_TYPE['pack'],
@@ -85,18 +51,14 @@ class Notifier(object):
             pack=ACTION_TRIGGER_TYPE['pack'],
             name=ACTION_TRIGGER_TYPE['name'])
 
-    def start(self):
-        self._consumer_thread = eventlet.spawn(self._queue_consumer.run)
-        self._consumer_thread.wait()
-
-    def handle_action_complete(self, liveaction):
+    def process(self, liveaction):
         if liveaction.status not in ACTION_COMPLETE_STATES:
-            LOG.exception('Received incorrect notification complete event. LiveAction=%s',
-                          liveaction)
+            LOG.warn('Received incorrect notification complete event. LiveAction=%s', liveaction)
             return
 
         if liveaction.notify is not None:
             self._post_notify_triggers(liveaction)
+
         self._post_generic_trigger(liveaction)
 
     def _post_notify_triggers(self, liveaction):
@@ -122,7 +84,12 @@ class Notifier(object):
             message = notify_subsection.message or (
                 'Action ' + liveaction.action + ' ' + default_message_suffix)
             data = notify_subsection.data or {}  # XXX: Handle Jinja
-            data = {'result': liveaction.result}
+            # At this point convert result to a string. This restricts the rulesengines
+            # ability to introspect the result. On the other handle atleast a json usable
+            # result is sent as part of the notification. If jinja is required to convert
+            # to a string representation it uses str(...) which make it impossible to
+            # parse the result as json any longer.
+            data['result'] = json.dumps(liveaction.result)
 
             payload['message'] = message
             payload['data'] = data
@@ -136,6 +103,8 @@ class Notifier(object):
             for channel in notify_subsection.channels:
                 try:
                     payload['channel'] = channel
+                    LOG.debug('POSTing %s for %s. Payload - %s.', NOTIFY_TRIGGER_TYPE['name'],
+                              liveaction.id, payload)
                     self._trigger_dispatcher.dispatch(self._notify_trigger, payload=payload)
                 except:
                     failed_channels.append(channel)
@@ -160,5 +129,4 @@ class Notifier(object):
 
 def get_notifier():
     with Connection(cfg.CONF.messaging.url) as conn:
-        notifier = Notifier(q_connection=conn, trigger_dispatcher=TriggerDispatcher(LOG))
-        return notifier
+        return Notifier(conn, [ACTIONUPDATE_WORK_Q], trigger_dispatcher=TriggerDispatcher(LOG))