+++ st2api/st2api/controllers/root.py
@@ -13,11 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-
 from pecan import expose
-from pecan.core import redirect
-from oslo.config import cfg
 
 from st2common import __version__
 from st2common import log as logging
@@ -25,24 +21,10 @@ import st2api.controllers.v1.root as v1_root
 import st2api.controllers.exp.root as exp_root
 
 __all__ = [
-    'WebUIRootController',
     'RootController'
 ]
 
 LOG = logging.getLogger(__name__)
-
-
-class WebUIRootController(object):
-    @expose(generic=True)
-    def index(self):
-        webui_root = os.path.join(cfg.CONF.api_pecan.static_root, 'webui')
-        webui_index_file = os.path.join(webui_root, 'index.html')
-
-        if not os.path.exists(webui_index_file):
-            message = 'Cannot find WebUI files. Make sure they are located in %s.' % (webui_root)
-            return message
-
-        return redirect(location='/webui/index.html')
 
 
 class RootController(object):
@@ -54,10 +36,6 @@ class RootController(object):
             'v1': v1_controller,
             'exp': exp_controller
         }
-
-        if cfg.CONF.api.serve_webui_files:
-            webui_controller = WebUIRootController()
-            self.controllers['webui'] = webui_controller
 
         self.default_controller = v1_controller
 
@@ -72,7 +50,6 @@ class RootController(object):
 
         data['version'] = __version__
         data['docs_url'] = docs_url
-        data['webui_enabled'] = cfg.CONF.api.serve_webui_files
         return data
 
     @expose()
@@ -80,9 +57,6 @@ class RootController(object):
         version = ''
         if len(remainder) > 0:
             version = remainder[0]
-            if cfg.CONF.api.serve_webui_files and version == 'webui':
-                # Special case for handling webui requests
-                return self.controllers['webui'], remainder[1:]
 
             if remainder[len(remainder) - 1] == '':
                 # If the url has a trailing '/' remainder will contain an empty string.