+++ st2api/tests/unit/controllers/exp/test_alias_execution.py
@@ -23,13 +23,13 @@ from tests import FunctionalTest
 FIXTURES_PACK = 'aliases'
 
 TEST_MODELS = {
-    'actionaliases': ['alias1.yaml', 'alias2.yaml'],
+    'aliases': ['alias1.yaml', 'alias2.yaml'],
     'actions': ['action1.yaml'],
     'runners': ['runner1.yaml']
 }
 
 TEST_LOAD_MODELS = {
-    'actionaliases': ['alias3.yaml']
+    'aliases': ['alias3.yaml']
 }
 
 
@@ -51,19 +51,22 @@ class TestAliasExecution(FunctionalTest):
         super(TestAliasExecution, cls).setUpClass()
         cls.models = FixturesLoader().save_fixtures_to_db(fixtures_pack=FIXTURES_PACK,
                                                           fixtures_dict=TEST_MODELS)
-        cls.alias1 = cls.models['actionaliases']['alias1.yaml']
-        cls.alias2 = cls.models['actionaliases']['alias2.yaml']
+        cls.alias1 = cls.models['aliases']['alias1.yaml']
+        cls.alias2 = cls.models['aliases']['alias2.yaml']
 
-    @mock.patch.object(action_service, 'schedule',
+    @mock.patch.object(action_service, 'request',
                        return_value=(None, DummyActionExecution(id_=1)))
-    def testBasicExecution(self, schedule):
+    def testBasicExecution(self, request):
         alias_execution = 'alias1 Lorem ipsum value1 dolor sit "value2 value3" amet.'
         post_resp = self._do_post(alias_execution)
         self.assertEqual(post_resp.status_int, 200)
         expected_parameters = {'param1': 'value1', 'param2': 'value2 value3'}
-        self.assertEquals(schedule.call_args[0][0].parameters, expected_parameters)
+        self.assertEquals(request.call_args[0][0].parameters, expected_parameters)
 
     def _do_post(self, execution, expect_errors=False):
-        execution = {'command': execution}
+        execution = {'command': execution,
+                     'user': 'stanley',
+                     'source_channel': 'test',
+                     'notification_channel': 'test'}
         return self.app.post_json('/exp/aliasexecution', execution,
                                   expect_errors=expect_errors)