+++ st2common/st2common/content/aliasesregistrar.py
@@ -13,6 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
+import six
+
 import st2common.content.utils as content_utils
 
 from st2common import log as logging
@@ -32,62 +35,116 @@ LOG = logging.getLogger(__name__)
 class AliasesRegistrar(ResourceRegistrar):
     ALLOWED_EXTENSIONS = ALLOWED_EXTS
 
-    def register_aliases_from_dirs(self, aliases_dirs=None):
-        if not aliases_dirs:
-            return 0
-        result = 0
-        for aliases_dir in aliases_dirs:
-            result += self.register_aliases_from_dir(aliases_dir=aliases_dir)
-        return result
+    def register_aliases_from_packs(self, base_dirs):
+        """
+        Discover all the packs in the provided directory and register aliases from all of the
+        discovered packs.
 
-    def register_aliases_from_dir(self, aliases_dir=None):
+        :return: Number of aliases registered.
+        :rtype: ``int``
+        """
+        registered_count = 0
+
+        content = self._pack_loader.get_content(base_dirs=base_dirs,
+                                                content_type='aliases')
+
+        for pack, aliases_dir in six.iteritems(content):
+            try:
+                LOG.debug('Registering aliases from pack %s:, dir: %s', pack, aliases_dir)
+                aliases = self._get_aliases_from_pack(aliases_dir)
+                count = self._register_aliases_from_pack(pack=pack, aliases=aliases)
+                registered_count += count
+            except:
+                LOG.exception('Failed registering all aliases from pack: %s', aliases_dir)
+
+        return registered_count
+
+    def register_aliases_from_pack(self, pack_dir):
+        """
+        Register all the aliases from the provided pack.
+
+        :return: Number of aliases registered.
+        :rtype: ``int``
+        """
+        pack_dir = pack_dir[:-1] if pack_dir.endswith('/') else pack_dir
+        _, pack = os.path.split(pack_dir)
+        aliases_dir = self._pack_loader.get_content_from_pack(pack_dir=pack_dir,
+                                                              content_type='aliases')
+
+        registered_count = 0
+
         if not aliases_dir:
-            return 0
-        resources = self._get_resources_from_pack(resources_dir=aliases_dir)
-        return self._register_aliases(aliases=resources)
+            return registered_count
 
-    def _register_aliases(self, aliases=None):
+        LOG.debug('Registering aliases from pack %s:, dir: %s', pack, aliases_dir)
+
+        try:
+            aliases = self._get_aliases_from_pack(aliases_dir=aliases_dir)
+            registered_count = self._register_aliases_from_pack(pack=pack, aliases=aliases)
+        except:
+            LOG.exception('Failed registering all aliases from pack: %s', aliases_dir)
+
+        return registered_count
+
+    def _get_aliases_from_pack(self, aliases_dir):
+        return self._get_resources_from_pack(resources_dir=aliases_dir)
+
+    def _register_action_alias(self, pack, action_alias):
+        content = self._meta_loader.load(action_alias)
+        pack_field = content.get('pack', None)
+        if not pack_field:
+            content['pack'] = pack
+            pack_field = pack
+        if pack_field != pack:
+            raise Exception('Model is in pack "%s" but field "pack" is different: %s' %
+                            (pack, pack_field))
+
+        action_alias_api = ActionAliasAPI(**content)
+        action_alias_api.validate()
+        action_alias_db = ActionAliasAPI.to_model(action_alias_api)
+
+        try:
+            action_alias_db.id = ActionAlias.get_by_name(action_alias_api.name).id
+        except ValueError:
+            LOG.info('ActionAlias %s not found. Creating new one.', action_alias)
+
+        try:
+            action_alias_db = ActionAlias.add_or_update(action_alias_db)
+            extra = {'action_alias_db': action_alias_db}
+            LOG.audit('Action alias updated. Action alias %s from %s.', action_alias_db,
+                      action_alias, extra=extra)
+        except Exception:
+            LOG.exception('Failed to create action alias %s.', action_alias_api.name)
+            raise
+
+    def _register_aliases_from_pack(self, pack, aliases):
         registered_count = 0
 
         for alias in aliases:
-            LOG.debug('Loading alias from %s.', alias)
             try:
-                content = self._meta_loader.load(alias)
-                action_alias_api = ActionAliasAPI(**content)
-                action_alias_db = ActionAliasAPI.to_model(action_alias_api)
-
-                try:
-                    action_alias_db.id = ActionAlias.get_by_name(action_alias_api.name).id
-                except ValueError:
-                    LOG.info('ActionAlias %s not found. Creating new one.', alias)
-
-                try:
-                    action_alias_db = ActionAlias.add_or_update(action_alias_db)
-                    extra = {'action_alias_db': action_alias_db}
-                    LOG.audit('Action alias updated. Action alias %s from %s.', action_alias_db,
-                              alias, extra=extra)
-                except Exception:
-                    LOG.exception('Failed to create action alias %s.', action_alias_api.name)
-
+                LOG.debug('Loading alias from %s.', alias)
+                self._register_action_alias(pack, alias)
             except Exception:
-                LOG.exception('Failed registering alias from %s.', alias)
+                LOG.exception('Unable to register alias: %s', alias)
+                continue
             else:
                 registered_count += 1
 
         return registered_count
 
 
-def register_aliases(aliases_base_paths=None, aliases_dir=None):
-    if aliases_base_paths:
-        assert(isinstance(aliases_base_paths, list))
+def register_aliases(packs_base_paths=None, pack_dir=None):
+    if packs_base_paths:
+        assert(isinstance(packs_base_paths, list))
+
+    if not packs_base_paths:
+        packs_base_paths = content_utils.get_packs_base_paths()
 
     registrar = AliasesRegistrar()
 
-    if aliases_dir:
-        result = registrar.register_aliases_from_dir(aliases_dir=aliases_dir)
+    if pack_dir:
+        result = registrar.register_aliases_from_pack(pack_dir=pack_dir)
     else:
-        if not aliases_base_paths:
-            aliases_base_paths = content_utils.get_aliases_base_paths()
-        result = registrar.register_aliases_from_dirs(aliases_dirs=aliases_base_paths)
+        result = registrar.register_aliases_from_packs(base_dirs=packs_base_paths)
 
     return result